Make map and filter more concurrent
Makes Promise.map, Promise.filter, .filter and .map more concurrent. Previously
they waited for the entire input array to be fulfilled before starting to
process it.

The callback order is still the same as before, e.g. if 4th promise in the array
fulfills while 2nd one is still pending, the callback for 4th value in the array
is not called until 2nd and 3rd are fulfilled and had their callbacks executed.