Handle promises arrays with properties
In case when variable "values" contains array of promises with properties (http://stackoverflow.com/questions/9952126/add-property-to-javascript-array) creation of new array causes the resolved array not to have the properties of original array (except length). Reusing the original array, created by bluebird user could resolve this issue and be a small optimization.