Handle promises arrays with properties
In case when variable "values" contains array of promises with properties (http://stackoverflow.com/questions/9952126/add-property-to-javascript-array) creation of new array causes the resolved array not to have the properties of original array (except length). One option to solve this would be reusing the original array, created by bluebird user. While it would be a small optimization this solution would have a funny side effect: turning the original array of promises, created by user to array of values.
Proposed solution is copiyng the custom properties from original array to the resolved array. This solution involves minimal performance overhead (http://jsperf.com/cloning-an-object). I do not use getOwnPropertyNames as it appeared in ECMA script 5.1 only.