Add a first draft of an Async Dialog Tutorial
This is an attempt to show Bluebird and Promises when used as values
from a dialog box. I use only Bluebird as a dependency and the DOM code
is vanilla JS not jQuery.

I first compare the `window.confirm` synchronous code with a
asynchronous dialog written in the DOM. Then I refactor the example to a
few OOP abstractions which allow be to demonstrate different types of
dialogs like a prompt.

I follow this with an explanation on how to best handle the cancellation
of the dialog box using custom Errors.

And finally I offer and example on how to provide progress updates by
using a progress bar dialog.