make tests work when run in a child_process
process.stdout.cursorTo and process.stdout.clearScreenDown only exist
if stdout is a tty. When smoke testing Node.js builds, citgm is used
to make sure breaking changes are not being made to popular packages.
This package fails due to not checking if the cursorTo and
clearScreenDown functions exists on process.stdout before using them.