doc: .finally handler returns a promise or throws
When the handler function passed to .finally throws an exception or
returns a promise which is later rejected, the promise returned by
.finally is rejected with the exception or rejected value.

Although this behavior does not appear to be currently documented, it
has been in place since at least 0.7.10.  It is also specified in both
Ron Buckton's[1] and Domenic Denicola's[2] proposals and present in all
of the implementations I checked[3][4][5][6].  So I assume it is safe to
document as part of the API.

This commit does just that.

1.  https://gist.github.com/rbuckton/66918c8491aa335b003c
2.  https://github.com/domenic/promises-unwrapping/issues/18
3.  https://github.com/cujojs/when/blob/3.7.7/lib/decorators/flow.js#L72
4.  https://github.com/tildeio/rsvp.js/blob/v3.2.1/lib/rsvp/promise.js#L437
5.  https://github.com/matthew-andrews/Promise.prototype.finally/blob/master/finally.js
6.  https://github.com/blakeembrey/promise-finally/blob/master/src/promise-finally.ts

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>