promise.resolve.md - static->promise example
Based this on the Promise.method. Not sure if that was a special case.

Also, not sure of my jQuery example description. It probably needs a clearer description in any case.
Is it just casting a jQuery then-able not-quite-promise to a Bluebird promise?
Special for jQuery or is this for any then-able object, say from another Promise library?