Update what-about-generators.md
No, this is not a complete article, but I wish this information would have been here earlier, so that I would have discovered the promises + generators combo a month ago. Please add it while waiting for a real article so others looking here might find it and benefit from it.