Fixed #1260: Add current line in creatorLog for debugging purpose
creatorLine take his lines from traceLines, but we have a variable "stack" that takes that traceLines and clean it (using cleanStack Fonction). 

In my case (you can check it on the issue), if we add to previous line the current line (with a verification in order to be sure that it exist as I'm not sure about why this has been done like that in first place). It fixes the problem. 
The fact that we do not delete anything avoid any regression (the first line will still be there), but add a new information (the second line) if it's available.