Add fallback for unhandled rejection timer on Firefox.
The default unhandled rejection timer is a setTimeout() from a removed
iframe. On Firefox, this throws NS_ERROR_NOT_INITIALIZED. This code
adds a test to verify that the setTimeout gleaned from the iframe is
functional and falls back to the global function if it isn’t.

This fixes support for rejecting promises on Firefox.

Fixes #1623.